refactor(components, app): move LargeButton to components and add helix styling (#15929)
closes AUTH-622
