+++ components/src/atoms/buttons/__tests__/LargeButton.test.tsx
@@ -2,9 +2,9 @@ import * as React from 'react'
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import '@testing-library/jest-dom/vitest'
 import { fireEvent, screen } from '@testing-library/react'
-import { COLORS } from '@opentrons/components'
-import { renderWithProviders } from '../../../__testing-utils__'
 
+import { renderWithProviders } from '../../../testing/utils'
+import { COLORS } from '../../../helix-design-system'
 import { LargeButton } from '../LargeButton'
 
 const render = (props: React.ComponentProps<typeof LargeButton>) => {
