+++ components/src/atoms/buttons/index.ts
@@ -1,4 +1,5 @@
 export * from './AlertPrimaryButton'
+export * from './LargeButton'
 export * from './PrimaryButton'
-export * from './SecondaryButton'
 export * from './RadioButton'
+export * from './SecondaryButton'
